Handle invoice payment by third party (kind of expense).
The third party payer (a known contact) must be defined on the invoice.
The invoice remains linked to the Vendor. 
When registering the payment, the Vendor invoice is marked as paid with
a matching move on an interim third party payment account using a thrid
party payment journal. 
The actual liquidity payment is immediately completed by a matching move
from this interim account to the liquidity account using the
corresponding liquidity journal.