FIX Allow group_logisfloo_admin to set the isshopreceipt flag.
A shop receipt was encoded and saved, but odoo reseted the flag to false
and the shop receipt was not finalised. The shop receipt was then
considered as a PO. Fixing this required to process the PO manually and
then to set the isshopreceipt to true so it appears in the right view.