Add Inventory period report and period management.
New report to provide consolidated stock level, sold and purchased values and margin for all product over a period of time.

A new InventoryPeriod object defines the start and end date of the period as well as the preceding period (as it is needed to compute the starting quantities).
Administrators can trigger the rebuild of an inventory period report.

Allow grouping and searching by vendor and category.

Tree view uses the following decorations:
- Mute archived product
- Alert on margin < 1
- Info for not calculated margins

Update Compute price history to use actual tax values.